bot/requirements.txt
hikari == 2.0.0.dev124 
hikari-lightbulb == 2.3.5
python-dotenv == 1.0.1
psycopg2 == 2.9.9